Move all files from ontology_building to tool_building
